% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boiwsa.R
\name{boiwsa}
\alias{boiwsa}
\title{Seasonal adjustment of weekly data}
\usage{
boiwsa(
  x,
  dates,
  r = 0.8,
  auto.ao.search = TRUE,
  out.threshold = 3.8,
  ao.list = NULL,
  my.k_l = NULL,
  H = NULL,
  ic = "aicc",
  method = "additive"
)
}
\arguments{
\item{x}{Input time series as a numeric vector}

\item{dates}{a vector of class "Date", containing the data dates}

\item{r}{Defines the rate of decay of the weights. Should be between zero and one. By default is set to 0.8.}

\item{auto.ao.search}{Boolean. Search for additive outliers}

\item{out.threshold}{t-stat threshold in outlier search. By default is 3.8}

\item{ao.list}{Vector with user specified additive outliers in a date format}

\item{my.k_l}{Numeric vector defining the number of yearly and monthly trigonometric variables. If NULL, is found automatically using the information criteria. The search range is 0:36 and 0:12 with the step size of 6 for the yearly and monthly variables, respectively.}

\item{H}{Matrix with holiday- and trading day factors}

\item{ic}{Information criterion used in the automatic search for the number of trigonometric regressors. There are thee options: aic, aicc and bic. By default uses aicc}

\item{method}{Decomposition type: additive or multiplicative}
}
\value{
sa Seasonally adjusted series

my.k_l Number of trigonometric variables used to model the seasonal pattern

sf Estimated seasonal effects

hol.factors Estimated holiday effects

out.factors Estimated outlier effects

beta Regression coefficients for the last year

m lm object. Unweighted OLS regression on the full sample
}
\description{
Performs seasonal adjustment and forecasting of weekly time series using a
regression-based decomposition framework estimated by discounted least squares.
Seasonality is modeled using trigonometric regressors, while holiday,
trading-day, and calendar effects are incorporated through additional
covariates. Additive outliers can be detected automatically using an optional
outlier search procedure. The function supports both additive and
multiplicative decompositions and produces seasonally adjusted series together
with the associated model components. Worked examples and additional usage
illustrations are available in the package repository on GitHub. See Ginker
(2024), \emph{boiwsa: An R Package for Seasonal Adjustment of Weekly Data},
\emph{The R Journal}, 16(3), 186--197, \doi{10.32614/RJ-2024-032}.
}
\details{
The methodological framework implemented in this function is described in
Ginker (2024), \emph{boiwsa: An R Package for Seasonal Adjustment of Weekly Data},
\emph{The R Journal}, 16(3), 186--197, \doi{10.32614/RJ-2024-032}.
}
\examples{
 # Not run
 # Seasonal adjustment of weekly US gasoline production
 \donttest{
 data("gasoline.data")
 res=boiwsa(x=gasoline.data$y,dates=gasoline.data$date)}
}
\author{
Tim Ginker
}
