% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dates_il}
\alias{dates_il}
\title{Israeli working dates}
\format{
A data frame with 21550 rows and 4 variables:
\describe{
\item{\code{DATE_VALUE}}{Date}
\item{\code{ISR_WORKING_DAY_PART}}{1: full working day, 0.5: half working day, 0: holiday}
\item{\code{JEWISH_FULL_DATE}}{Jewish date}
\item{\code{DATE_WEEK_NUMBER}}{Weekday}
}
}
\source{
Personal
}
\usage{
dates_il
}
\description{
Israeli working dates
}
\keyword{datasets}
