% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find additive outliers}
\usage{
find_outliers(
  y,
  dates,
  out.tolerance = 3.8,
  my.AO.list = NULL,
  H = NULL,
  my.k_l = NULL
)
}
\arguments{
\item{y}{Numeric vector with a detrended dependent variable}

\item{dates}{Vector with dates in a date format}

\item{out.tolerance}{t-stat threshold for outliers (see Findley et al., 1998)}

\item{my.AO.list}{Vector with user defined additive outlier variables}

\item{H}{Matrix with holiday and trading day variables}

\item{my.k_l}{Vector with the number of fourier terms to capture the yeraly and monthly cycle. If NULL, would perform automatic search}
}
\value{
my.k_l

ao list of AO dates
}
\description{
Find additive outliers
}
