% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasoline.data}
\alias{gasoline.data}
\title{US finished motor gasoline product supplied}
\format{
\subsection{Data.Frame}{

A data frame with 1355 rows and 2 columns:
\describe{
\item{date}{date in a date format}
\item{y}{gasoline consumption}
}
}
}
\source{
fpp2 package
}
\usage{
gasoline.data
}
\description{
Weekly data beginning 2 February 1991, ending 20 January 2017. Units are "million barrels per day".
Report ...
}
\keyword{datasets}
