% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiday_dates_il}
\alias{holiday_dates_il}
\title{Israeli moving holiday dates}
\format{
A data frame with 51 rows and 3 variables:
\describe{
\item{\code{year}}{Year}
\item{\code{rosh}}{Rosh Hashanah date}
\item{\code{pesah}}{Pesach date}
}
}
\source{
Personal
}
\usage{
holiday_dates_il
}
\description{
Rosh Hashanah and Pesach dates
}
\keyword{datasets}
