% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ao.R
\name{my_ao}
\alias{my_ao}
\title{Create additive outlier indicator variables}
\usage{
my_ao(dates, out.list)
}
\arguments{
\item{dates}{A vector of class \code{"Date"} corresponding to the observation dates.}

\item{out.list}{A vector of class \code{"Date"} specifying candidate additive
outlier dates.}
}
\value{
A numeric matrix with \code{length(dates)} rows and one column per
outlier date present in \code{dates}. Column names are of the form
\code{"AO <date>"}. If none of the supplied outlier dates coincide with
\code{dates}, \code{NULL} is returned.
}
\description{
Constructs a matrix of additive outlier (AO) indicator variables based on a
set of user-specified outlier dates. For each outlier date that coincides with
an observation date, a binary indicator equal to one is created at the
corresponding position and zero elsewhere. Outlier dates not present in the
observation dates are silently ignored.
}
\examples{

# create a sequence of dates
dates=seq.Date(from=as.Date("2023-01-02"),by="weeks",length.out = 100)
# create a vector of outlier dates
my_ao_dates=as.Date(c("2023-01-02","2023-01-03"))
# create a matrix of AO variables
my_ao(dates = dates,out.list = my_ao_dates)
# as you can see there is only one column corresponding to 2023-01-02,
# the second date is ignored because it is not present in the dates vector

}
