% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.boiwsa}
\alias{plot.boiwsa}
\title{Plot}
\usage{
\method{plot}{boiwsa}(x, ...)
}
\arguments{
\item{x}{Result of boiwsa}

\item{...}{Additional arguments (currently not used).}
}
\description{
S3 method for objects of class "boiwsa".
Produces a ggplot object of seasonally decomposed time series.
}
