% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print.boiwsa}
\alias{print.boiwsa}
\title{Print method for boiwsa objects}
\usage{
\method{print}{boiwsa}(x, ...)
}
\arguments{
\item{x}{Result of \code{boiwsa}.}

\item{...}{Additional arguments (currently not used).}
}
\description{
S3 method for objects of class \code{boiwsa}. Prints a short model summary
including the number of trigonometric terms and the position of outliers.
}
