% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print}
\alias{print}
\title{Generic print function}
\usage{
print(x, ...)
}
\arguments{
\item{x}{An object to print.}

\item{...}{Additional arguments (currently not used).}
}
\description{
This is the generic print function.
}
