% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{rankUpdateInverse}
\alias{rankUpdateInverse}
\title{Update the inverse of a cross product of a matrix X when adding a new column v}
\usage{
rankUpdateInverse(X_inv, X_t, v)
}
\arguments{
\item{X_inv}{The inverse (X^T X)^-1 before adding the new column}

\item{X_t}{The transpose of X, i.e. X^T}

\item{v}{The column to add}
}
\value{
The inverse of (\link{X v}^T \link{X v})^-1
}
\description{
Update the inverse of a cross product of a matrix X when adding a new column v
}
