% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{summary.boiwsa}
\alias{summary.boiwsa}
\title{Summary function}
\usage{
\method{summary}{boiwsa}(object, ...)
}
\arguments{
\item{object}{An object of class \code{boiwsa}.}

\item{...}{Additional arguments (currently not used).}
}
\description{
S3 method for objects of class "boiwsa". Prints the regression summary output.
}
