% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{summary}
\alias{summary}
\title{Generic summary function}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object to summarize.}

\item{...}{Additional arguments (currently not used).}
}
\description{
This is the generic summary function.
}
