% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boiwsa.R
\name{boiwsa}
\alias{boiwsa}
\title{Weekly Seasonal adjustment}
\usage{
boiwsa(
  x,
  dates,
  r = 0.8,
  auto.ao.seacrh = T,
  ao.list = NULL,
  my.k_l = NULL,
  H = NULL,
  ic = "aicc",
  method = "additive"
)
}
\arguments{
\item{method}{}
}
\description{
Weekly Seasonal adjustment
}
