% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_opt.R
\name{find_opt}
\alias{find_opt}
\title{Find optimal number of fourier variables}
\usage{
find_opt(y, dates, H = NULL, AO = NULL)
}
\arguments{
\item{y}{Numeric vector with a detrended dependent variable}

\item{dates}{Vector with dates in a date format}

\item{H}{Matrix with holiday and trading day variables}

\item{AO}{Matrix with additive outlier variables}
}
\value{
list with the optimal number of fourier terms according to AIC, AICc and BIC
}
\description{
Find optimal number of fourier variables
}
