% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_vars.R
\name{fourier_vars}
\alias{fourier_vars}
\title{Create fourier predictors}
\usage{
fourier_vars(k = 1, l = 1, dates)
}
\arguments{
\item{k}{Number of yearly cycle fourier terms}

\item{l}{Number of monthly cycle fourier terms}

\item{dates}{Vector of dates in a date format}
}
\value{
Matrix with fourier variables
}
\description{
Creates sine and cosine variables to capture intramonthly and intrayearly cycles.
}
\examples{

# create a vector of dates
dates=seq.Date(from=as.Date("2023-01-02"),by="weeks",length.out = 100)
# Create a matrix with 20 yearly and 6 monthly pairs of sine and cosine variables
X=fourier_vars(k=20,l=6,dates=dates)

}
