% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_vars.R
\name{fourier_vars}
\alias{fourier_vars}
\title{Create fourier predictors}
\usage{
fourier_vars(k = 1, l = 1, dates)
}
\arguments{
\item{k}{Number of yearly cycle fourier terms}

\item{l}{Number of monthly cycle fourier terms}

\item{dates}{Vector of dates in a date format}
}
\value{
Matrix with fourier variables
}
\description{
Create fourier predictors
}
