% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genhol.R
\name{genhol}
\alias{genhol}
\title{Generate Holiday Regression Variables}
\usage{
genhol(dates, holiday.dates, start = 7, end = 7)
}
\arguments{
\item{dates}{a vector of class "Date", containing the data dates}

\item{holiday.dates}{a vector of class "Date", containing the occurrences of the holiday. It can be generated with as.Date().}

\item{start}{integer, shifts the start point of the holiday. Use negative values if start is before the specified date.}

\item{end}{integer, shifts end point of the holiday. Use negative values if end is before the specified date.}
}
\value{
an matrix with holiday variables that can be used as a user defined variable in boiwsa().
}
\description{
Generate Holiday Regression Variables
}
\examples{

# Creating moving holiday variable for Israeli Rosh Hashanah
data(gasoline.data)
movehol=genhol(gasoline.data$date,holiday.dates = holiday_dates_il$rosh)


}
