% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbm}
\alias{lbm}
\title{Initial unemployment claims in Israel (adjusted for strikes)}
\format{
\subsection{Data.Frame}{

A data frame with 313 rows and 2 columns:
\describe{
\item{date}{date in a date format}
\item{IES_IN_W_ADJ}{Initial unemployment claims}
}
}
}
\source{
Internal
}
\usage{
lbm
}
\description{
Weekly data beginning 11 January 2014, ending 4 January 2020.
}
\keyword{datasets}
