% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ao.R
\name{my_ao}
\alias{my_ao}
\title{Create additive outlier variables}
\usage{
my_ao(dates, out.list)
}
\arguments{
\item{dates}{Vector of dates in a date format}

\item{out.list}{Vector of outlier dates in a date format}
}
\value{
AO matrix with outlier variables
}
\description{
Create additive outlier variables
}
