% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_td.R
\name{simple_td}
\alias{simple_td}
\title{Generate simple trading day variable}
\usage{
simple_td(dates, df.td)
}
\arguments{
\item{dates}{a vector of class "Date", containing the data dates}

\item{df.td}{dataframe with working days. See, for example, dates_il}
}
\value{
matrix with trading day variables
}
\description{
Generate simple trading day variable
}
